% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetICA_extract_model.R
\name{MetICA_extract_model}
\alias{MetICA_extract_model}
\title{Extracting MetICA model}
\usage{
MetICA_extract_model(M1, ics, tops = ics)
}
\arguments{
\item{M1}{The entire list object generated from the function MetICA(X, pcs = 15...)}

\item{ics}{Number of clusters (MetICA components) decided by users according to validationPlot}

\item{tops}{Integer or NULL object. If integer, the tops "best" components based on each evaluation criterion will be plotted.}
}
\value{
MetICA model (a list object) whose number of components is decided by users.
\itemize{
 \item{"S"}{Scores of MetICA components. Data matrix contains n rows (samples) and ics columns (components)}
 \item{"A"}{Loadings of MetICA components. Data matrix contains p rows (metabolic features) and ics columns (components)}
 \item{"eval$kurtosis"}{Vector of length ics containing Kurtosis measure of each MetICA component.}
 \item{"eval$cluster_size"}{Number of IPCA estimates in each corresponding cluster.}
 \item{"eval$divergence_index"}{Divergence indexes of clusters.}
 \item{"eval$boot_prop"}{Proportion of bootstrap estimates in each cluster.}
 \item{"eval$boot_stability"}{Stability of each cluster against bootstrapping.}
}
}
\description{
This function extracts a MetICA model with user-decided component numbers.
}
