% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_metabolome.R
\docType{data}
\name{yeast_metabolome}
\alias{yeast_metabolome}
\title{Yeast exo-metabolome example data}
\format{A two-element list object:
\itemize{
  \item features: metabolic features detected and extracted. Each feature is represented by a unique ID and a mass (m/z).
  \item X: data matrix with 45 rows (samples) and 2700 columns (metabolic features). This is the example data matrix for MetICA simulations.
}}
\usage{
data(yeast_metabolome)
}
\description{
Exo-metabolomic profiles of 15 yeast commercial strains for winemaking. The experiment included 3 biological replicates. Samples were taken at the end of alcoholic fermentationand measured on direct-infusion FT-ICR-MS.
}
\references{
Youzhong Liu, Sara Forcisi, Mourad Harir, Magali Deleris-Bou, Sibylle Krieger-Weber, Marianna Lucio, Cedric Longin, Claudine Degueurce, Regis D. Gougeon, Philippe Schmitt-Kopplin and Herve Alexandre, New molecular evidence of wine yeast-bacteria interaction unraveled by non-targeted exometabolomic profiling, Metabolomics (2016) Vol. 12 no. 69
}
\keyword{datasets}
